requirements.txt
opencv-python>=4.12.0.88
onnxruntime-gpu>=1.22.0
numba>=0.61.2
insightface>=0.7.3
scikit-learn>=1.7.2
scikit-image>=0.25.2
deep-sort-realtime>=1.3.2
albumentationsx>=2.0.11
